test: 스타카토 정보 불러오는 기능에서 토큰 검증 추가
